Add lambda handlers
This commit also does some cleanup amongst other things...
